Add files via upload
Hi, sir/madam this is my first online hackathon and so i hope you understand that it may have some flaws. i mainly did the code by my own but took help from Chatgpt. I couldn't the backend with the frontend as i am in second year and i did not have the time to learn it. I hope you understand this.